pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "asimov-cli"
version = "1.0.0"
description = "Advanced LoRA Fine-tuning CLI for Vast.ai with comprehensive job management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Asimov CLI Team", email = "contact@asimov-cli.com"}
]
maintainers = [
    {name = "Asimov CLI Team", email = "contact@asimov-cli.com"}
]
keywords = ["machine-learning", "fine-tuning", "lora", "vast-ai", "cli", "pytorch", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.7"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pathlib2>=2.3.0;python_version<'3.4'",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.8",
    "setuptools>=45",
    "wheel>=0.37",
    "twine>=3.4",
    "build>=0.7.0",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/asimov-cli"
Documentation = "https://github.com/yourusername/asimov-cli#readme"
Repository = "https://github.com/yourusername/asimov-cli.git"
"Bug Tracker" = "https://github.com/yourusername/asimov-cli/issues"
Changelog = "https://github.com/yourusername/asimov-cli/blob/main/CHANGELOG.md"

[project.scripts]
asimov = "asimov_cli:cli"

[tool.setuptools]
py-modules = ["asimov_cli", "vast_manager", "main"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = "E203,W503"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"